requirements-dev.txt
pytest-django==4.8.0
pluggy==1.5.0
pytest==8.2.2
packaging==24.1
iniconfig==2.0.0
Faker==26.0.0
python-dateutil==2.9.0.post0
factory-boy==3.3.0
six==1.16.0
